Planning/src/pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "claude-agent-sdk>=0.1.40",
    "anyio",
    "pyyaml",
    "python-dotenv",
    "postgrest",
    "openai>=2.20",
    "openai-codex-sdk>=0.1",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
